update only resource props recursively
