+++ localstack/services/cloudformation/cloudformation_api.py
@@ -164,15 +164,6 @@ class Stack(object):
             ("LogicalResourceId", resource_id),
             ("PhysicalResourceId", physical_res_id),
         )
-        print(
-            "!!!!_set_resource_status_details",
-            resource_id,
-            id(resource),
-            physical_res_id,
-            attr_defaults,
-            resource.get("PhysicalResourceId"),
-            state.get("PhysicalResourceId"),
-        )
         for res in [resource, state]:
             for attr, default in attr_defaults:
                 res[attr] = res.get(attr) or default
@@ -180,13 +171,14 @@ class Stack(object):
         state["StackId"] = state.get("StackId") or self.stack_id
         state["ResourceType"] = state.get("ResourceType") or self.resources[resource_id].get("Type")
 
-        print(
-            "!!!!_set_resource_status_details AFTER",
-            resource_id,
-            id(resource),
-            resource.get("PhysicalResourceId"),
-            state.get("PhysicalResourceId"),
-        )
+        if resource.get("Type") == "AWS::EC2::Instance":
+            print(
+                "!!!!_set_resource_status_details AFTER",
+                resource_id,
+                id(resource),
+                resource.get("PhysicalResourceId"),
+                state.get("PhysicalResourceId"),
+            )
 
     def resource_status(self, resource_id: str):
         result = self._lookup(self.resource_states, resource_id)
